requirements.txt
torch
torchvision
timm
numpy
tqdm
Pillow